src/FastAPI_Server(1)/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
langchain
langchain-community
huggingface-hub
torch
transformers      